Update CEMO013 and CEMO014
